+++ .phan/config.php
@@ -7,6 +7,7 @@ $cfg['directory_list'] = array_merge(
 	[
 		'../../extensions/CategoryTree',
 		'../../extensions/Disambiguator',
+		'../../extensions/GrowthExperiments',
 		'../../extensions/PageViewInfo',
 		'../../extensions/Wikibase',
 	]
@@ -17,6 +18,7 @@ $cfg['exclude_analysis_directory_list'] = array_merge(
 	[
 		'../../extensions/CategoryTree',
 		'../../extensions/Disambiguator',
+		'../../extensions/GrowthExperiments',
 		'../../extensions/PageViewInfo',
 		'../../extensions/Wikibase',
 	]